requirements.txt
pandas
matplotlib
seaborn
langchain-community
langchain-chroma
python-dotenv
gradio
langchain-openai
transformers
kagglehub